data/requirements.txt
numpy
matplotlib
h5py
pyyaml
tqdm